Implement SKIPTO environment variable (action #497)

+++ inst/startqemu.pm
@@ -62,7 +62,7 @@ if($ENV{UPGRADE} && !$ENV{LIVECD}) {
 	}
 }
 
-if(!$ENV{KEEPHDDS}) {
+if(!$ENV{KEEPHDDS} && !$ENV{SKIPTO}) {
 	# fresh HDDs
 	for my $i (1..4) {
 		unlink("$basedir/l$i");
